use commander to parse cli arguments
add `commander` dependency. use that to parse the cli arguments, instead
of using `process.argv` directly

currently only the format command (as before) and use that as default.
this allows formatting with just the file name, or with `format
filename` and will allow easier adding more subcommands in the future